- renamed: RestaurantRepository -> RestaurantService VoteRepository -> VoteService - fixed create() and update() methods

+++ src/main/java/graduation/graduationProject/service/VoteService.java
@@ -4,34 +4,35 @@ import graduation.graduationProject.model.Vote;
 import graduation.graduationProject.repository.CrudRestaurantRepository;
 import graduation.graduationProject.repository.CrudUserRepository;
 import graduation.graduationProject.repository.CrudVoteRepository;
-import graduation.graduationProject.util.exception.ApplicationException;
-import org.springframework.stereotype.Repository;
+import graduation.graduationProject.util.exception.ModificationRestrictionException;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.List;
 
 import static graduation.graduationProject.util.ValidationUtil.checkNotFoundWithId;
 
-@Repository
-public class VoteRepository {
+@Service
+public class VoteService {
 
     private final CrudVoteRepository crudVoteRepository;
     private final CrudUserRepository crudUserRepository;
     private final CrudRestaurantRepository crudRestaurantRepository;
 
-    public VoteRepository(CrudVoteRepository crudVoteRepository,
-                          CrudUserRepository crudUserRepository,
-                          CrudRestaurantRepository crudRestaurantRepository) {
+    public VoteService(CrudVoteRepository crudVoteRepository,
+                       CrudUserRepository crudUserRepository,
+                       CrudRestaurantRepository crudRestaurantRepository) {
         this.crudVoteRepository = crudVoteRepository;
         this.crudUserRepository = crudUserRepository;
         this.crudRestaurantRepository = crudRestaurantRepository;
     }
 
     @Transactional
-    public Vote save(int userId, int restaurant_id) {
-        if (get(userId) != null){
-            throw new ApplicationException("You cannot change your vote today");
+    public Vote create(int userId, int restaurant_id) {
+        if (isAfterEleven() && get(userId) != null) {
+            throw new ModificationRestrictionException();
         }
         Vote vote = new Vote();
         vote.setUser(crudUserRepository.getOne(userId));
@@ -52,7 +53,16 @@ public class VoteRepository {
     }
 
     public boolean delete(int id, int userId) {
+        if (isAfterEleven()) {
+            throw new ModificationRestrictionException();
+        }
         checkNotFoundWithId(crudVoteRepository.delete(id, userId) != 0, id);
         return true;
     }
+
+    private boolean isAfterEleven() {
+        return LocalTime.now().isAfter(LocalTime.of(11, 0));
+    }
+
+
 }
